+++ src/main/java/seedu/hustler/command/task/ByeCommand.java
@@ -1,4 +1,4 @@
-package seedu.hustler.command.taskCommand;
+package seedu.hustler.command.task;
 
 import javafx.application.Platform;
 import seedu.hustler.command.Command;
