+++ src/main/java/seedu/hustler/command/task/UndoCommand.java
@@ -1,4 +1,4 @@
-package seedu.hustler.command.taskCommand;
+package seedu.hustler.command.task;
 
 import seedu.hustler.command.Command;
 import seedu.hustler.data.MemoryManager;
